ft_printf

+++ format/ft_printf_format.h
@@ -6,7 +6,7 @@
 /*   By: kipark <kipark@student.42seoul.kr>         +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/01/21 19:07:00 by kipark            #+#    #+#             */
-/*   Updated: 2022/01/24 19:58:11 by kipark           ###   ########seoul.kr  */
+/*   Updated: 2022/01/30 21:08:51 by kipark           ###   ########seoul.kr  */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -19,13 +19,13 @@
 
 # define HEXADECIMAL_MAX_LEN 16
 
-int	ft_printf_format_c(va_list ap);
-int	ft_printf_format_s(va_list ap);
-int	ft_printf_format_d(va_list ap);
-int	ft_printf_format_i(va_list ap);
-int	ft_printf_format_u(va_list ap);
-int	ft_printf_format_x(va_list ap);
-int	ft_printf_format_x_x(va_list ap);
-int	ft_printf_format_p(va_list ap);
+ssize_t	ft_printf_format_c(va_list ap);
+ssize_t	ft_printf_format_s(va_list ap);
+ssize_t	ft_printf_format_d(va_list ap);
+ssize_t	ft_printf_format_i(va_list ap);
+ssize_t	ft_printf_format_u(va_list ap);
+ssize_t	ft_printf_format_x(va_list ap);
+ssize_t	ft_printf_format_x_x(va_list ap);
+ssize_t	ft_printf_format_p(va_list ap);
 
 #endif
