+++ format/ft_printf_format_i.c
@@ -6,13 +6,13 @@
 /*   By: kipark <kipark@student.42seoul.kr>         +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/01/21 19:05:13 by kipark            #+#    #+#             */
-/*   Updated: 2022/01/24 19:54:48 by kipark           ###   ########seoul.kr  */
+/*   Updated: 2022/01/30 18:00:08 by kipark           ###   ########seoul.kr  */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include"ft_printf_format.h"
 
-int	ft_printf_format_i(va_list ap)
+ssize_t	ft_printf_format_i(va_list ap)
 {
 	return (ft_printf_format_d(ap));
 }
