+++ format/ft_printf_format_s.c
@@ -6,13 +6,13 @@
 /*   By: kipark <kipark@student.42seoul.kr>         +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/01/21 19:05:13 by kipark            #+#    #+#             */
-/*   Updated: 2022/01/24 19:55:01 by kipark           ###   ########seoul.kr  */
+/*   Updated: 2022/01/30 18:00:00 by kipark           ###   ########seoul.kr  */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include"ft_printf_format.h"
 
-int	ft_printf_format_s(va_list ap)
+ssize_t	ft_printf_format_s(va_list ap)
 {
 	return (ft_printf_format_write(va_arg(ap, char *)));
 }
