+++ format/ft_printf_format_x_x.c
@@ -6,7 +6,7 @@
 /*   By: kipark <kipark@student.42seoul.kr>         +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/01/21 19:05:13 by kipark            #+#    #+#             */
-/*   Updated: 2022/01/24 20:02:11 by kipark           ###   ########seoul.kr  */
+/*   Updated: 2022/01/30 17:59:53 by kipark           ###   ########seoul.kr  */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -27,7 +27,7 @@ static int	ft_printf_recursive_x_x(unsigned int result)
 	return (write_byte);
 }
 
-int	ft_printf_format_x_x(va_list ap)
+ssize_t	ft_printf_format_x_x(va_list ap)
 {
 	return (ft_printf_recursive_x_x(va_arg(ap, unsigned int)));
 }
